backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
Pillow>=10.0.0
tenacity>=8.0.0
aiofiles>=23.0.0
httpx>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.0
bcrypt>=4.0.0,<5.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
eval-type-backport>=0.2.0
